Add more prefab shapes, improve type hinting
Added 4 new custom shapes to be used as prefabs. All can be found inside shapes/prefabs.

Also finally imported three.js types. This greatly improves the type inference throughout the whole program.

Also made it that the small spheres found on vertices do not scale with the rest of the shape as this was leading to weird looking outcomes.