Expand on very basic UI
Added very basic modal for shape properties and added basic sidebar. Sidebar is mainly for testing how canvas reacts when it is not the only thing on the screen. Modal currently has no styling, just shows basic properties. Will expand on properties and add a settings menu to allow for changing to be made to selected shape.